Remove favorite from user and deck so it's stored in the database
In 98bf5aea we moved the Favorite relationship owner to deck, which
means that we need to modify it in the deck to store the change in the
database. In this commit we also added the favorite to the deck on
saving, but not on removing. This commit fixes it, so it also is removed
properly.